Changes name or constants.h -> rnp_defs.h and introduces rnp_types.h
