+++ src/lib/compress.c
@@ -75,7 +75,7 @@ __RCSID("$NetBSD: compress.c,v 1.23 2012/03/05 02:20:18 christos Exp $");
 #include <string.h>
 #include <stdlib.h>
 
-#include "constants.h"
+#include "rnp_defs.h"
 #include "packet-parse.h"
 #include "errors.h"
 #include "rnpdefs.h"
