+++ src/lib/eddsa.c
@@ -29,7 +29,7 @@
 #include <string.h>
 #include <botan/ffi.h>
 #include "utils.h"
-#include "constants.h"
+#include "rnp_defs.h"
 
 int
 pgp_genkey_eddsa(pgp_seckey_t *seckey, size_t curve_len)
