+++ src/lib/packet-create.c
@@ -95,7 +95,7 @@ __RCSID("$NetBSD: create.c,v 1.38 2010/11/15 08:03:39 agc Exp $");
 #include "pgp-key.h"
 #include "rnpsdk.h"
 #include "ecdsa.h"
-#include "constants.h"
+#include "rnp_defs.h"
 
 extern ec_curve_desc_t ec_curves[PGP_CURVE_MAX];
